requirements.txt
requests==2.31.0
python-dotenv==1.0.0
colorama==0.4.6
pytest==7.4.0

